Async and Helper Unification
Dependecy Injecting the HttpClient now. Converted everything to use
async. Instead of HttpClientUtil now everything uses a DI IHttpClient.
Created BlockchainApiHelper that groups all the other helpers together
to be used

+++ src/Info.Blockchain.API/BlockExplorer/BitcoinValue.cs
@@ -18,7 +18,7 @@
 
 		public decimal Bits => this.btc * BitcoinValue.BitsPerBitcoin;
 
-		public decimal Satoshis => this.btc * BitcoinValue.SatoshisPerBitcoin;
+		public long Satoshis => (long)this.btc * BitcoinValue.SatoshisPerBitcoin;
 
 		public static BitcoinValue Zero => new BitcoinValue();
